+++ crates/ink/codegen/src/generator/as_dependency/contract_ref.rs
@@ -89,6 +89,33 @@ impl ContractRef<'_> {
         let storage_ident = self.contract.module().storage().ident();
         let ref_ident = self.generate_contract_ref_ident();
         let abi = default_abi!();
+        let sol_codec = if cfg!(any(ink_abi = "sol", ink_abi = "all")) {
+            // TODO: (@davidsemakula) Replace with derived implementations when available.
+            quote_spanned!(span=>
+                impl<Abi> ::ink::SolDecode for #ref_ident<Abi> {
+                    type SolType = ::ink::Address;
+
+                    fn from_sol_type(value: Self::SolType) -> Self {
+                        Self {
+                            inner: <<#storage_ident
+                                as ::ink::codegen::ContractCallBuilder>::Type<Abi>
+                                as ::ink::env::call::FromAddr>::from_addr(value),
+                            _marker: ::core::default::Default::default(),
+                        }
+                    }
+                }
+
+                impl<'a, Abi> ::ink::SolEncode<'a> for #ref_ident<Abi> {
+                    type SolType = &'a ::ink::Address;
+
+                    fn to_sol_type(&'a self) -> Self::SolType {
+                        self.as_ref()
+                    }
+                }
+            )
+        } else {
+            quote!()
+        };
         quote_spanned!(span=>
             #[derive(
                 ::core::fmt::Debug,
@@ -182,6 +209,8 @@ impl ContractRef<'_> {
                         >::type_info()
                     }
                 }
+
+                #sol_codec
             };
         )
     }