+++ crates/ink/codegen/src/generator/trait_def/call_forwarder.rs
@@ -160,6 +160,38 @@ impl CallForwarder<'_> {
     fn generate_auxiliary_trait_impls(&self) -> TokenStream2 {
         let span = self.span();
         let call_forwarder_ident = self.ident();
+        let sol_codec = if cfg!(any(ink_abi = "sol", ink_abi = "all")) {
+            // TODO: (@davidsemakula) Replace with derived implementations when available.
+            quote_spanned!(span=>
+                impl<E, Abi> ::ink::SolDecode for #call_forwarder_ident<E, Abi>
+                where
+                    E: ::ink::env::Environment,
+                {
+                    type SolType = ::ink::Address;
+
+                    fn from_sol_type(value: Self::SolType) -> Self {
+                        Self {
+                            builder: <<Self as ::ink::codegen::TraitCallBuilder>::Builder
+                                as ::ink::env::call::FromAddr>::from_addr(value),
+                            _marker: ::core::default::Default::default(),
+                        }
+                    }
+                }
+
+                impl<'a, E, Abi> ::ink::SolEncode<'a> for #call_forwarder_ident<E, Abi>
+                where
+                    E: ::ink::env::Environment,
+                {
+                    type SolType = &'a ::ink::Address;
+
+                    fn to_sol_type(&'a self) -> Self::SolType {
+                        self.as_ref()
+                    }
+                }
+            )
+        } else {
+            quote!()
+        };
         quote_spanned!(span=>
             /// We require this manual implementation since the derive produces incorrect trait bounds.
             impl<E, Abi> ::core::clone::Clone for #call_forwarder_ident<E, Abi>
@@ -207,6 +239,8 @@ impl CallForwarder<'_> {
                     >::type_info()
                 }
             }
+
+            #sol_codec
         )
     }