+++ crates/ink/codegen/src/generator/trait_def/message_builder.rs
@@ -109,6 +109,34 @@ impl MessageBuilder<'_> {
     fn generate_auxiliary_trait_impls(&self) -> TokenStream2 {
         let span = self.span();
         let message_builder_ident = self.trait_def.message_builder_ident();
+        let sol_codec = if cfg!(any(ink_abi = "sol", ink_abi = "all")) {
+            // TODO: (@davidsemakula) Replace with derived implementations when available.
+            quote_spanned!(span=>
+                impl<E, Abi> ::ink::SolDecode for #message_builder_ident<E, Abi>
+                where
+                    E: ::ink::env::Environment,
+                {
+                    type SolType = ();
+
+                    fn from_sol_type(_: Self::SolType) -> Self {
+                        Self {
+                            _marker: ::core::default::Default::default(),
+                        }
+                    }
+                }
+
+                impl<'a, E, Abi> ::ink::SolEncode<'a> for #message_builder_ident<E, Abi>
+                where
+                    E: ::ink::env::Environment,
+                {
+                    type SolType = ();
+
+                    fn to_sol_type(&'a self) {}
+                }
+            )
+        } else {
+            quote!()
+        };
         quote_spanned!(span=>
             impl<E, Abi> ::core::default::Default for #message_builder_ident<E, Abi>
             where
@@ -127,6 +155,8 @@ impl MessageBuilder<'_> {
             {
                 type Env = E;
             }
+
+            #sol_codec
         )
     }