+++ crates/ink/tests/ui/abi/sol/pass/constructor-contract-ref.rs
@@ -1,0 +1,25 @@
+#![allow(unexpected_cfgs)]
+
+#[ink::contract]
+mod contract {
+    #[ink(storage)]
+    pub struct Contract {
+        contract: Option<ContractRef>
+    }
+
+    impl Contract {
+        #[ink(constructor)]
+        pub fn constructor() -> Self {
+            Self {
+                contract: None
+            }
+        }
+
+        #[ink(message)]
+        pub fn message(&mut self, contract: ContractRef) {
+            self.contract = Some(contract);
+        }
+    }
+}
+
+fn main() {}